+++ cli/docs/changelog.md
@@ -2,6 +2,10 @@
 ### Changes
 * Replaced `EnclaveManager` with `Kurtosis Engine API Libs` which handle all the interactions with the `Kurtosis Engine Server`
 
+### Features
+* Added documentation in README about how to develop on this repo
+* Upgraded to `kurtosis-core` 1.25.2, which contains fixes for `container-engine-lib` 0.7.0 feature that allows binding container ports to specific host machine ports
+
 ### Fixes
 * Clarified the difference between the two types of params in `module exec`